requirements.txt
tqdm
opencv-python
numpy
torch
torchvision
logging
matplotlib
pillow